+++ src/flatObject/extractList.js
@@ -7,6 +7,7 @@ return function extractList ( list, options ) {
             const 
                       root     = indexFn ( 'root' )
                     , { main:{load}, INIT_DATA_TYPES } = dependencies ()
+                    , asTypes = [ ...INIT_DATA_TYPES, 'dt-object' ]
                     ;
             let 
                   error = false
@@ -18,9 +19,9 @@ return function extractList ( list, options ) {
                             error = true
                             errorMsg = `Options should be an object and property "as" is required`
                         }
-                    if ( !error && !INIT_DATA_TYPES.includes(options.as) ) {  
+                    if ( !error && !asTypes.includes(options.as) ) {  
                             error = true
-                            errorMsg = `Invalid option "as" value: ${options.as}. Choose one of: std, standard, ''`
+                            errorMsg = `Invalid option "as" value: ${options.as}.`
                         }
                     if ( error )  throw new Error ( errorMsg )
                 }
@@ -34,7 +35,7 @@ return function extractList ( list, options ) {
                         .map ( item => {
                                             if ( item == null            )   return null
                                             if ( !(item instanceof Array))   return item
-                                            return options ? load(item).model(() => options ) : load(item)
+                                            return load(item).model(() => options )
                                         })
 }} // extractList func.